# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Secure account foundation API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RegisterInput {
  /**
     * @minLength 2
     * @maxLength 80
     */
  name: string;
  email: string;
  /**
     * @minLength 8
     * @maxLength 128
     */
  password: string;
}

export interface LoginInput {
  email: string;
  /**
     * @minLength 1
     * @maxLength 128
     */
  password: string;
}

export interface EmailInput {
  email: string;
}

export interface EmailTokenInput {
  /** @minLength 20 */
  token: string;
}

export interface PasswordResetInput {
  /** @minLength 20 */
  token: string;
  /**
     * @minLength 8
     * @maxLength 128
     */
  password: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  user: 'user',
  admin: 'admin',
} as const;

export interface User {
  id: string;
  name: string;
  email: string;
  role: UserRole;
  emailVerified: boolean;
  createdAt: string;
}

export type AuthResponseTokenType = typeof AuthResponseTokenType[keyof typeof AuthResponseTokenType];


export const AuthResponseTokenType = {
  bearer: 'bearer',
} as const;

export interface AuthResponse {
  accessToken: string;
  tokenType: AuthResponseTokenType;
  user: User;
}

export interface MessageResponse {
  message: string;
}

export interface ErrorResponse {
  error: string;
}

/**
 * Bad request
 */
export type BadRequestResponse = ErrorResponse;

/**
 * Authentication required
 */
export type UnauthorizedResponse = ErrorResponse;

/**
 * Insufficient permissions
 */
export type ForbiddenResponse = ErrorResponse;

/**
 * Resource already exists
 */
export type ConflictResponse = ErrorResponse;