# lib/api-zod/src/generated/types/authResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Secure account foundation API
 * OpenAPI spec version: 0.1.0
 */
import type { AuthResponseTokenType } from './authResponseTokenType';
import type { User } from './user';

export interface AuthResponse {
  accessToken: string;
  tokenType: AuthResponseTokenType;
  user: User;
}