# lib/api-zod/src/generated/types/emailTokenInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Secure account foundation API
 * OpenAPI spec version: 0.1.0
 */

export interface EmailTokenInput {
  /** @minLength 20 */
  token: string;
}