# lib/api-zod/src/generated/types/loginInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Secure account foundation API
 * OpenAPI spec version: 0.1.0
 */

export interface LoginInput {
  email: string;
  /**
     * @minLength 1
     * @maxLength 128
     */
  password: string;
}