# lib/api-zod/src/generated/types/passwordResetInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Secure account foundation API
 * OpenAPI spec version: 0.1.0
 */

export interface PasswordResetInput {
  /** @minLength 20 */
  token: string;
  /**
     * @minLength 8
     * @maxLength 128
     */
  password: string;
}